requirements.txt
pywebview
Pillow
torch
numpy
tqdm
safetensors
scipy
scikit-learn
transformers
